+++ merger/app/merger/logging.go
@@ -18,4 +18,4 @@ import (
 	"github.com/streamingfast/logging"
 )
 
-var zlog, tracer = logging.PackageLogger("merger", "github.com/streamingfast/firehose-core/merger/app/merger")
+var zlog, _ = logging.PackageLogger("merger", "github.com/streamingfast/firehose-core/merger/app/merger")