+++ merger/bundler.go
@@ -15,11 +15,9 @@
 package merger
 
 import (
-	"bytes"
 	"context"
 	"errors"
 	"fmt"
-	"io"
 	"sync"
 	"time"
 
@@ -29,6 +27,7 @@ import (
 	"github.com/streamingfast/bstream/forkable"
 	"github.com/streamingfast/firehose-core/merger/metrics"
 	"github.com/streamingfast/logging"
+	"go.uber.org/atomic"
 	"go.uber.org/zap"
 )
 
@@ -53,6 +52,9 @@ type Bundler struct {
 	irreversibleBlocks []*bstream.OneBlockFile
 	forkable           *forkable.Forkable
 
+	blockTimestampInFlight *atomic.Bool
+	blockTimestampLastRun  *atomic.Int64 // unix nanos
+
 	logger *zap.Logger
 }
 
@@ -60,13 +62,15 @@ var logger, _ = logging.PackageLogger("merger", "github.com/streamingfast/fireho
 
 func NewBundler(startBlock, stopBlock, firstStreamableBlock, bundleSize uint64, io IOInterface) *Bundler {
 	b := &Bundler{
-		bundleSize:           bundleSize,
-		io:                   io,
-		bundleError:          make(chan error, 1),
-		firstStreamableBlock: firstStreamableBlock,
-		stopBlock:            stopBlock,
-		seenBlockFiles:       make(map[string]*bstream.OneBlockFile),
-		logger:               logger,
+		bundleSize:             bundleSize,
+		io:                     io,
+		bundleError:            make(chan error, 1),
+		firstStreamableBlock:   firstStreamableBlock,
+		stopBlock:              stopBlock,
+		seenBlockFiles:         make(map[string]*bstream.OneBlockFile),
+		blockTimestampInFlight: atomic.NewBool(false),
+		blockTimestampLastRun:  atomic.NewInt64(0),
+		logger:                 logger,
 	}
 	b.Reset(toBaseNum(startBlock, bundleSize), nil)
 	return b
@@ -126,19 +130,6 @@ func (b *Bundler) Reset(nextBase uint64, lib bstream.BlockRef) {
 	b.Unlock()
 }
 
-func readBlockTime(data []byte) (time.Time, error) {
-	reader := bytes.NewReader(data)
-	blockReader, err := bstream.NewDBinBlockReader(reader)
-	if err != nil {
-		return time.Time{}, fmt.Errorf("unable to create block reader: %w", err)
-	}
-	blk, err := blockReader.Read()
-	if err != nil && err != io.EOF {
-		return time.Time{}, fmt.Errorf("block reader failed: %w", err)
-	}
-	return blk.Time(), nil
-}
-
 func (b *Bundler) ProcessBlock(_ *pbbstream.Block, obj interface{}) error {
 	obf := obj.(bstream.ObjectWrapper).WrappedObject().(*bstream.OneBlockFile)
 	if obf.Num < b.baseBlockNum {
@@ -169,18 +160,19 @@ func (b *Bundler) ProcessBlock(_ *pbbstream.Block, obj interface{}) error {
 		metrics.AppReadiness.SetReady()
 		b.irreversibleBlocks = append(b.irreversibleBlocks, obf)
 		metrics.HeadBlockNumber.SetUint64(obf.Num)
-		go func() {
-			// this pre-downloads the data
-			data, err := obf.Data(context.Background(), b.io.DownloadOneBlockFile)
-			if err != nil {
-				return
-			}
-			// now that we have the data, might as well read the block time for metrics
-			if time, err := readBlockTime(data); err == nil {
-				metrics.HeadBlockTimeDrift.SetBlockTime(time)
-			}
-		}()
 		b.Unlock()
+		if time.Since(time.Unix(0, b.blockTimestampLastRun.Load())) >= time.Second*5 && b.blockTimestampInFlight.CompareAndSwap(false, true) {
+			b.blockTimestampLastRun.Store(time.Now().UnixNano())
+			go func() {
+				defer b.blockTimestampInFlight.Store(false)
+				t, err := readBlockTimestamp(context.Background(), obf, b.io.OpenOneBlockFile)
+				if err != nil {
+					b.logger.Debug("cannot read block timestamp for head drift metric", zap.Error(err))
+					return
+				}
+				metrics.HeadBlockTimeDrift.SetBlockTime(t)
+			}()
+		}
 		return nil
 	}